# src/docs/swagger.yaml
definitions:
  dto.CountryResponse:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  dto.CreateUpdateCountryRequest:
    properties:
      name:
        maxLength: 32
        minLength: 3
        type: string
    required:
    - name
    type: object
  dto.GetOtpRequest:
    properties:
      mobileNumber:
        maxLength: 11
        minLength: 11
        type: string
    required:
    - mobileNumber
    type: object
  dto.LoginByUsernameRequest:
    properties:
      password:
        minLength: 6
        type: string
      username:
        minLength: 5
        type: string
    required:
    - password
    - username
    type: object
  dto.RegisterLoginByMobileRequest:
    properties:
      mobileNumber:
        maxLength: 11
        minLength: 11
        type: string
      otp:
        maxLength: 6
        minLength: 6
        type: string
    required:
    - mobileNumber
    - otp
    type: object
  dto.RegisterUserByUsernameRequest:
    properties:
      email:
        minLength: 6
        type: string
      firstName:
        minLength: 3
        type: string
      lastName:
        minLength: 6
        type: string
      password:
        minLength: 6
        type: string
      username:
        minLength: 5
        type: string
    required:
    - firstName
    - lastName
    - password
    - username
    type: object
  handlers.PersonData:
    properties:
      first_name:
        maxLength: 20
        minLength: 3
        type: string
      last_name:
        maxLength: 20
        minLength: 3
        type: string
      mobile_number:
        maxLength: 11
        minLength: 11
        type: string
    required:
    - first_name
    - last_name
    - mobile_number
    type: object
  helper.BaseHttpResponse:
    properties:
      error: {}
      result: {}
      resultCode:
        $ref: '#/definitions/helper.ResultCode'
      success:
        type: boolean
      validationError:
        items:
          $ref: '#/definitions/validations.ValidationError'
        type: array
    type: object
  helper.ResultCode:
    enum:
    - 0
    - 40001
    - 40101
    - 40301
    - 40401
    - 42901
    - 42902
    - 50001
    - 50002
    type: integer
    x-enum-varnames:
    - Success
    - ValidationError
    - AuthError
    - ForbiddenError
    - NotFoundError
    - LimiterError
    - OtpLimiterError
    - CustomRecovery
    - InternalError
  validations.ValidationError:
    properties:
      message:
        type: string
      property:
        type: string
      tag:
        type: string
      value:
        type: string
    type: object
info:
  contact: {}
paths:
  /v1/countries:
    post:
      consumes:
      - application/json
      description: Create a country
      parameters:
      - description: create a country
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUpdateCountryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Country response
          schema:
            allOf:
            - $ref: '#/definitions/helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/dto.CountryResponse'
              type: object
        "400":
          description: Bad request
          schema:
            allOf:
            - $ref: '#/definitions/helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/dto.CountryResponse'
              type: object
      security:
      - AuthBearer: []
      summary: Create a country
      tags:
      - Countries
  /v1/countries/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a country
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: response
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Delete a country
      tags:
      - Countries
    get:
      consumes:
      - application/json
      description: Get a country
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: country response
          schema:
            allOf:
            - $ref: '#/definitions/helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/dto.CountryResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Get a country
      tags:
      - Countries
    put:
      consumes:
      - application/json
      description: Update a country
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: update a country
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUpdateCountryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Country response
          schema:
            allOf:
            - $ref: '#/definitions/helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/dto.CountryResponse'
              type: object
        "400":
          description: Bad request
          schema:
            allOf:
            - $ref: '#/definitions/helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/dto.CountryResponse'
              type: object
      security:
      - AuthBearer: []
      summary: Update a country
      tags:
      - Countries
  /v1/health/:
    get:
      consumes:
      - application/json
      description: Health Check
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
      summary: Health Check
      tags:
      - health
  /v1/test/binder/body:
    post:
      consumes:
      - application/json
      description: User Id Des
      parameters:
      - description: person data
        in: body
        name: person
        required: true
        schema:
          $ref: '#/definitions/handlers.PersonData'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: User ID
      tags:
      - test
  /v1/test/binder/uri/{id}/{name}:
    get:
      consumes:
      - application/json
      description: User Id Des
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: integer
      - description: user name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
      summary: User ID
      tags:
      - test
  /v1/users/login-by-mobile:
    post:
      consumes:
      - application/json
      description: RegisterLoginByMobileNumber
      parameters:
      - description: RegisterLoginByMobileRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterLoginByMobileRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Success
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
        "409":
          description: Failed
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
      summary: RegisterLoginByMobileNumber
      tags:
      - Users
  /v1/users/login-by-username:
    post:
      consumes:
      - application/json
      description: LoginByUsername
      parameters:
      - description: LoginByUsernameRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginByUsernameRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Success
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
        "409":
          description: Failed
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
      summary: LoginByUsername
      tags:
      - Users
  /v1/users/register-by-username:
    post:
      consumes:
      - application/json
      description: RegisterByUsername
      parameters:
      - description: RegisterUserByUsernameRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterUserByUsernameRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Success
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
        "409":
          description: Failed
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
      summary: RegisterByUsername
      tags:
      - Users
  /v1/users/send-otp:
    post:
      consumes:
      - application/json
      description: Send otp to user
      parameters:
      - description: get otp request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.GetOtpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Success
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
        "409":
          description: Failed
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
      summary: Send otp to user
      tags:
      - Users
securityDefinitions:
  AuthBearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"